DSX V3 API + small DSX V2 API Update
