+++ xchange-examples/src/main/java/org/knowm/xchange/examples/dsx/DSXExamplesUtils.java
@@ -9,9 +9,9 @@ import org.knowm.xchange.dsx.DSXExchange;
 /** @author Mikhail Wall */
 public class DSXExamplesUtils {
 
-  public static Exchange createExchange() throws IOException {
+  public static Exchange createExchange(Class<? extends Exchange> exchangeClass) throws IOException {
 
-    ExchangeSpecification exSpec = new ExchangeSpecification(DSXExchange.class);
+    ExchangeSpecification exSpec = new ExchangeSpecification(exchangeClass);
     exSpec.setSecretKey("6FY66EC0YJ7YL4OHAC1X5ZEYAUUFRJBFI21I9S7A5PW7");
     exSpec.setApiKey("396f12b3-9998-4ff8-800b-f29ab6d8e886");
     exSpec.setSslUri("https://dsx.uk");
