+++ xchange-examples/src/main/java/org/knowm/xchange/examples/dsx/marketdata/DSXDepthDemoV3.java
@@ -1,22 +1,23 @@
 package org.knowm.xchange.examples.dsx.marketdata;
 
-import java.io.IOException;
-import java.util.Map;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeFactory;
 import org.knowm.xchange.currency.CurrencyPair;
-import org.knowm.xchange.dsx.DSXExchange;
+import org.knowm.xchange.dsx.DSXExchangeV3;
 import org.knowm.xchange.dsx.dto.marketdata.DSXOrderbook;
-import org.knowm.xchange.dsx.service.DSXMarketDataService;
+import org.knowm.xchange.dsx.service.DSXMarketDataServiceV3;
 import org.knowm.xchange.dto.marketdata.OrderBook;
 import org.knowm.xchange.service.marketdata.MarketDataService;
 
-/** @author Mikhail Wall */
-public class DSXDepthDemo {
+import java.io.IOException;
+import java.util.Map;
+
+/** @author Pavel Chertalev */
+public class DSXDepthDemoV3 {
 
   public static void main(String[] args) throws IOException {
 
-    Exchange dsx = ExchangeFactory.INSTANCE.createExchange(DSXExchange.class.getName());
+    Exchange dsx = ExchangeFactory.INSTANCE.createExchange(DSXExchangeV3.class);
     generic(dsx);
     raw(dsx);
   }
@@ -40,8 +41,8 @@ public class DSXDepthDemo {
 
   private static void raw(Exchange exchange) throws IOException {
 
-    DSXMarketDataService marketDataService =
-        (DSXMarketDataService) exchange.getMarketDataService();
+    DSXMarketDataServiceV3 marketDataService =
+        (DSXMarketDataServiceV3) exchange.getMarketDataService();
 
     Map<String, DSXOrderbook> depth =
         marketDataService.getDSXOrderbook("ltcusd", "LIVE").getOrderbookMap();
