+++ xchange-examples/src/main/java/org/knowm/xchange/examples/dsx/marketdata/DSXTickerDemoV3.java
@@ -1,24 +1,25 @@
 package org.knowm.xchange.examples.dsx.marketdata;
 
-import java.io.IOException;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeFactory;
 import org.knowm.xchange.currency.CurrencyPair;
-import org.knowm.xchange.dsx.DSXExchange;
+import org.knowm.xchange.dsx.DSXExchangeV3;
 import org.knowm.xchange.dsx.dto.marketdata.DSXTickerWrapper;
-import org.knowm.xchange.dsx.service.DSXMarketDataService;
+import org.knowm.xchange.dsx.service.DSXMarketDataServiceV3;
 import org.knowm.xchange.dto.marketdata.Ticker;
 import org.knowm.xchange.service.marketdata.MarketDataService;
 import org.knowm.xchange.utils.CertHelper;
 
+import java.io.IOException;
+
 /** @author Mikhail Wall */
-public class DSXTickerDemo {
+public class DSXTickerDemoV3 {
 
   public static void main(String[] args) throws Exception {
 
     CertHelper.trustAllCerts();
 
-    Exchange dsx = ExchangeFactory.INSTANCE.createExchange(DSXExchange.class.getName());
+    Exchange dsx = ExchangeFactory.INSTANCE.createExchange(DSXExchangeV3.class);
     generic(dsx);
     raw(dsx);
   }
@@ -39,8 +40,8 @@ public class DSXTickerDemo {
 
   private static void raw(Exchange exchange) throws IOException {
 
-    DSXMarketDataService marketDataService =
-        (DSXMarketDataService) exchange.getMarketDataService();
+    DSXMarketDataServiceV3 marketDataService =
+        (DSXMarketDataServiceV3) exchange.getMarketDataService();
 
     DSXTickerWrapper ticker = marketDataService.getDSXTicker("btcusd", "LIVE");
     System.out.println(ticker.toString());
