+++ xchange-examples/src/main/java/org/knowm/xchange/examples/dsx/marketdata/DSXTradesDemoV3.java
@@ -1,22 +1,23 @@
 package org.knowm.xchange.examples.dsx.marketdata;
 
-import java.io.IOException;
-import java.util.Map;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeFactory;
 import org.knowm.xchange.currency.CurrencyPair;
-import org.knowm.xchange.dsx.DSXExchange;
+import org.knowm.xchange.dsx.DSXExchangeV3;
 import org.knowm.xchange.dsx.dto.marketdata.DSXTrade;
-import org.knowm.xchange.dsx.service.DSXMarketDataService;
+import org.knowm.xchange.dsx.service.DSXMarketDataServiceV3;
 import org.knowm.xchange.dto.marketdata.Trades;
 import org.knowm.xchange.service.marketdata.MarketDataService;
 
+import java.io.IOException;
+import java.util.Map;
+
 /** @author Mikhail Wall */
-public class DSXTradesDemo {
+public class DSXTradesDemoV3 {
 
   public static void main(String[] args) throws IOException {
 
-    Exchange dsx = ExchangeFactory.INSTANCE.createExchange(DSXExchange.class.getName());
+    Exchange dsx = ExchangeFactory.INSTANCE.createExchange(DSXExchangeV3.class);
     generic(dsx);
     raw(dsx);
   }
@@ -32,8 +33,8 @@ public class DSXTradesDemo {
 
   private static void raw(Exchange exchange) throws IOException {
 
-    DSXMarketDataService marketDataService =
-        (DSXMarketDataService) exchange.getMarketDataService();
+    DSXMarketDataServiceV3 marketDataService =
+        (DSXMarketDataServiceV3) exchange.getMarketDataService();
 
     Map<String, DSXTrade[]> trades =
         marketDataService.getDSXTrades("btcusd", 7, "LIVE").getTradesMap();
