+++ xchange-examples/src/main/java/org/knowm/xchange/examples/dsx/trade/DSXUserTransHistoryDemoV3.java
@@ -1,29 +1,30 @@
 package org.knowm.xchange.examples.dsx.trade;
 
-import java.io.IOException;
-import java.util.Map;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.dsx.DSXExchange;
+import org.knowm.xchange.dsx.DSXExchangeV3;
 import org.knowm.xchange.dsx.dto.trade.DSXOrder;
 import org.knowm.xchange.dsx.dto.trade.DSXOrderHistoryResult;
 import org.knowm.xchange.dsx.dto.trade.DSXTransHistoryResult;
-import org.knowm.xchange.dsx.service.DSXTradeService;
+import org.knowm.xchange.dsx.service.DSXTradeServiceV3;
 import org.knowm.xchange.examples.dsx.DSXExamplesUtils;
 import org.knowm.xchange.exceptions.ExchangeException;
 
+import java.io.IOException;
+import java.util.Map;
+
 /** @author Mikhail Wall */
-public class DSXUserTransHistoryDemo {
+public class DSXUserTransHistoryDemoV3 {
 
   public static void main(String[] args) throws IOException {
 
-    Exchange dsx = DSXExamplesUtils.createExchange(DSXExchange.class);
+    Exchange dsx = DSXExamplesUtils.createExchange(DSXExchangeV3.class);
     reaw(dsx);
     raw(dsx);
     raww(dsx);
   }
 
   private static void raw(Exchange exchange) throws IOException {
-    DSXTradeService tradeService = (DSXTradeService) exchange.getTradeService();
+    DSXTradeServiceV3 tradeService = (DSXTradeServiceV3) exchange.getTradeService();
     Map<Long, DSXTransHistoryResult> trans = null;
     try {
       trans = tradeService.getDSXTransHistory(null, null, null, null, null, null, null, null, null);
@@ -37,7 +38,7 @@ public class DSXUserTransHistoryDemo {
   }
 
   private static void raww(Exchange exchange) throws IOException {
-    DSXTradeService tradeService = (DSXTradeService) exchange.getTradeService();
+    DSXTradeServiceV3 tradeService = (DSXTradeServiceV3) exchange.getTradeService();
     Map<Long, DSXOrderHistoryResult> orders = null;
     try {
       orders = tradeService.getDSXOrderHistory(null, null, null, null, null, null, null);
@@ -51,7 +52,7 @@ public class DSXUserTransHistoryDemo {
   }
 
   private static void reaw(Exchange exchange) throws IOException {
-    DSXTradeService tradeService = (DSXTradeService) exchange.getTradeService();
+    DSXTradeServiceV3 tradeService = (DSXTradeServiceV3) exchange.getTradeService();
     Map<Long, DSXOrder> orders = null;
     try {
       orders = tradeService.getDSXActiveOrders("btcusd");
